+++ eslint.config.js
@@ -60,6 +60,7 @@ module.exports = [
     ignores: [
       "js/vendor/**",
       "js/data/xlsx-templates.js",
+      "js/data/drum-clips.js",
       "node_modules/**",
       "faq/screenshots/**",
       "coverage/**",